Intermediate event story commit.
- Add DI support
- Delete useless logs
- Do small refactoring

// More to come

- Add UI
