+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/use_cases/EmployeeInfoUseCase.kt
@@ -1,12 +1,11 @@
-package band.effective.office.tv.useCases
+package band.effective.office.tv.network.use_cases
 
 import band.effective.office.tv.domain.models.Employee.EmployeeInfoEntity
 import band.effective.office.tv.repository.EmployeeInfoRepositoryImpl
 import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
 
-class EmployeeInfoUseCase(
-    private val repositoryImpl: EmployeeInfoRepositoryImpl
-) {
+class EmployeeInfoUseCase @Inject constructor(private val repositoryImpl: EmployeeInfoRepositoryImpl) {
     fun getLatestBirthdays(): Flow<List<EmployeeInfoEntity>> {
         return repositoryImpl.latestBirthdays
     }
