+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/view/viewmodel/EventStoryViewModel.kt
@@ -1,6 +1,5 @@
 package band.effective.office.tv.view.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import band.effective.office.tv.domain.LatestEventInfoUiState
@@ -8,20 +7,23 @@ import band.effective.office.tv.domain.models.Employee.Anniversary
 import band.effective.office.tv.domain.models.Employee.Birthday
 import band.effective.office.tv.domain.models.Employee.EmployeeInfo
 import band.effective.office.tv.domain.models.Employee.NewEmployee
-import band.effective.office.tv.useCases.EmployeeInfoUseCase
+import band.effective.office.tv.network.use_cases.EmployeeInfoUseCase
 import band.effective.office.tv.utils.DateUtlils.getYearsWithTheCompany
 import band.effective.office.tv.utils.DateUtlils.isCelebrationToday
 import band.effective.office.tv.utils.DateUtlils.isNewEmployeeToday
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.catch
+import javax.inject.Inject
 
-class EventStoryViewModel(
-    private val employeeInfoUseCase: EmployeeInfoUseCase,
-) : ViewModel() {
+@HiltViewModel
+class EventStoryViewModel @Inject constructor(private var employeeInfoUseCase: EmployeeInfoUseCase) :
+    ViewModel() {
     val uiState =
         MutableStateFlow<LatestEventInfoUiState>(LatestEventInfoUiState.Success(emptyList()))
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
+
 
     init {
         fetchBirthdays()
@@ -35,7 +37,6 @@ class EventStoryViewModel(
                 }.collect { birthdays ->
                     val resultList = mutableListOf<EmployeeInfo>()
                     birthdays.map {
-                        Log.d("BirthdayScreenViewModel", it.startDate)
                         if (isCelebrationToday(it.nextBirthdayDate)) {
                             resultList.add(
                                 Birthday(
